requirements.txt
Flask
Pillow
numpy
sklearn
